# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LecturerUploadInput {
  file: Blob;
}

export interface ErrorResponse {
  error: string;
}

export type LecturerJobStatus = typeof LecturerJobStatus[keyof typeof LecturerJobStatus];


export const LecturerJobStatus = {
  pending: 'pending',
  extracting: 'extracting',
  analyzing: 'analyzing',
  generating: 'generating',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface LecturerJob {
  id: string;
  status: LecturerJobStatus;
  progressStep?: string | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  progressPct?: number | null;
  inputFilename: string;
  extractedTextCount?: number | null;
  extractedImageCount?: number | null;
  slideCount?: number | null;
  error?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface SlideImage {
  originalIndex: number;
  dataBase64?: string | null;
  mimeType?: string | null;
  altText?: string | null;
}

export type SlideDataType = typeof SlideDataType[keyof typeof SlideDataType];


export const SlideDataType = {
  title: 'title',
  section_header: 'section_header',
  content: 'content',
  data_table: 'data_table',
  chart: 'chart',
  comparison: 'comparison',
  callout: 'callout',
} as const;

export type SlideDataChartType = typeof SlideDataChartType[keyof typeof SlideDataChartType] | null;


export const SlideDataChartType = {
  bar: 'bar',
  line: 'line',
  pie: 'pie',
} as const;

export type SlideDataChartData = { [key: string]: unknown } | null;

export type SlideDataCalloutStyle = typeof SlideDataCalloutStyle[keyof typeof SlideDataCalloutStyle] | null;


export const SlideDataCalloutStyle = {
  definition: 'definition',
  warning: 'warning',
  takeaway: 'takeaway',
} as const;

export interface SlideData {
  index: number;
  type: SlideDataType;
  title?: string | null;
  subtitle?: string | null;
  body?: string | null;
  images?: SlideImage[];
  tableHeaders?: string[] | null;
  tableRows?: string[][] | null;
  chartType?: SlideDataChartType;
  chartData?: SlideDataChartData;
  leftColumn?: string | null;
  rightColumn?: string | null;
  calloutIcon?: string | null;
  calloutStyle?: SlideDataCalloutStyle;
  textBlockIndices?: number[];
}

export interface IntegrityReport {
  textBlocksExtracted: number;
  textBlocksPlaced: number;
  imagesExtracted: number;
  imagesPlaced: number;
  unplacedTextIndices: number[];
  unplacedImageIndices: number[];
  allPlaced: boolean;
}

export interface SlidesResult {
  slides: SlideData[];
  integrity: IntegrityReport;
}