# lib/api-zod/src/generated/types/integrityReport.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface IntegrityReport {
  textBlocksExtracted: number;
  textBlocksPlaced: number;
  imagesExtracted: number;
  imagesPlaced: number;
  unplacedTextIndices: number[];
  unplacedImageIndices: number[];
  allPlaced: boolean;
}