# lib/api-zod/src/generated/types/lecturerJob.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LecturerJobStatus } from './lecturerJobStatus';

export interface LecturerJob {
  id: string;
  status: LecturerJobStatus;
  progressStep?: string | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  progressPct?: number | null;
  inputFilename: string;
  extractedTextCount?: number | null;
  extractedImageCount?: number | null;
  slideCount?: number | null;
  error?: string | null;
  createdAt: Date;
  updatedAt: Date;
}