# lib/api-zod/src/generated/types/slideData.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SlideDataCalloutStyle } from './slideDataCalloutStyle';
import type { SlideDataChartData } from './slideDataChartData';
import type { SlideDataChartType } from './slideDataChartType';
import type { SlideDataType } from './slideDataType';
import type { SlideImage } from './slideImage';

export interface SlideData {
  index: number;
  type: SlideDataType;
  title?: string | null;
  subtitle?: string | null;
  body?: string | null;
  images?: SlideImage[];
  tableHeaders?: string[] | null;
  tableRows?: string[][] | null;
  chartType?: SlideDataChartType;
  chartData?: SlideDataChartData;
  leftColumn?: string | null;
  rightColumn?: string | null;
  calloutIcon?: string | null;
  calloutStyle?: SlideDataCalloutStyle;
  textBlockIndices?: number[];
}