# lib/api-zod/src/generated/types/slideImage.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SlideImage {
  originalIndex: number;
  dataBase64?: string | null;
  mimeType?: string | null;
  altText?: string | null;
}