# lib/api-zod/src/generated/types/slidesResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { IntegrityReport } from './integrityReport';
import type { SlideData } from './slideData';

export interface SlidesResult {
  slides: SlideData[];
  integrity: IntegrityReport;
}